refactor: shared-metadata and structured-data
