+++ apps/web/src/app/(landing)/blog/category/[slug]/page.tsx
@@ -1,11 +1,11 @@
 import { ContentCategory } from "@/app/(landing)/content-category";
 import { ContentList } from "@/app/(landing)/content-list";
+import { getBlogPosts } from "@/content/utils";
 import {
   defaultMetadata,
   ogMetadata,
   twitterMetadata,
-} from "@/app/shared-metadata";
-import { getBlogPosts } from "@/content/utils";
+} from "@/lib/metadata/shared-metadata";
 import type { Metadata } from "next";
 
 const TITLE = "Blog Category";
