+++ apps/web/src/app/(landing)/play/checker/[slug]/page.tsx
@@ -1,15 +1,12 @@
 import {
-  BASE_URL,
-  getJsonLDBreadcrumbList,
-  getPageMetadata,
-} from "@/app/shared-metadata";
-import {
   getCheckerDataById,
   latencyFormatter,
   regionFormatter,
 } from "@/components/ping-response-analysis/utils";
 import { CustomMDX } from "@/content/mdx";
 import { getToolsPage } from "@/content/utils";
+import { BASE_URL, getPageMetadata } from "@/lib/metadata/shared-metadata";
+import { getJsonLDBreadcrumbList } from "@/lib/metadata/structured-data";
 import type { Metadata } from "next";
 import { redirect } from "next/navigation";
 import type { BreadcrumbList, WithContext } from "schema-dts";
