+++ apps/web/src/app/(landing)/play/uptime-sla/page.tsx
@@ -1,10 +1,7 @@
-import {
-  BASE_URL,
-  getJsonLDBreadcrumbList,
-  getPageMetadata,
-} from "@/app/shared-metadata";
 import { CustomMDX } from "@/content/mdx";
 import { getToolsPage } from "@/content/utils";
+import { BASE_URL, getPageMetadata } from "@/lib/metadata/shared-metadata";
+import { getJsonLDBreadcrumbList } from "@/lib/metadata/structured-data";
 import type { Metadata } from "next";
 import type { BreadcrumbList, WithContext } from "schema-dts";
 import { Calculation } from "./client";
