+++ apps/web/src/app/api/og/monitor/route.tsx
@@ -2,8 +2,8 @@ import { ImageResponse } from "next/og";
 
 import { OSTinybird } from "@openstatus/tinybird";
 
-import { DESCRIPTION, TITLE } from "@/app/shared-metadata";
 import { env } from "@/env";
+import { DESCRIPTION, TITLE } from "@/lib/metadata/shared-metadata";
 import { BasicLayout } from "../_components/basic-layout";
 import { Tracker } from "../_components/tracker";
 import { SIZE, calSemiBold, interLight, interRegular } from "../utils";
