+++ apps/web/src/app/api/og/route.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @next/next/no-img-element */
 import { ImageResponse } from "next/og";
 
-import { OG_DESCRIPTION, TITLE } from "@/app/shared-metadata";
+import { OG_DESCRIPTION, TITLE } from "@/lib/metadata/shared-metadata";
 import { SIZE } from "./utils";
 
 export const runtime = "edge";
